refactor response deletion

+++ plugins/kick_deleted_plugin.py
@@ -35,8 +35,9 @@ async def kick_deleted(event):
         except errors.UserAdminInvalidError:
             response = await event.respond("UserAdminInvalidError:  "
                                             + "One of the admins has deleted their account, so I cannot kick it from the group.")
-    await sleep(60)
-    await response.delete()
+    if response:
+        await sleep(60)
+        await response.delete()
 
     if kicked_users < 0:
         return
